chore: rename old wizard
