+++ src/lib/layout/index.ts
@@ -8,7 +8,7 @@ export { default as Notification } from './notification.svelte';
 export { default as Notifications } from './notifications.svelte';
 export { default as Shell } from './shell.svelte';
 export { default as Logs } from './logs.svelte';
-export { default as Wizard } from './wizard.svelte';
+export { default as Wizard } from './wizardWithSteps.svelte';
 export { default as WizardStep } from './wizardStep.svelte';
 export { default as Breadcrumbs } from './breadcrumbs.svelte';
 export { default as Unauthenticated } from './unauthenticated.svelte';
