+++ src/lib/wizards/functions/connectExisting.svelte
@@ -2,7 +2,7 @@
     import { Wizard } from '$lib/layout';
     import SelectRepository from './steps/selectRepository.svelte';
     import GitConfiguration from './steps/gitConfiguration.svelte';
-    import type { WizardStepsType } from '$lib/layout/wizard.svelte';
+    import type { WizardStepsType } from '$lib/layout/wizardWithSteps.svelte';
     import { sdk } from '$lib/stores/sdk';
     import { func } from '$routes/(console)/project-[project]/functions/function-[function]/store';
     import { choices, installation, repository } from './store';
