+++ src/routes/(console)/project-[project]/functions/create-function/wizard/connectExisting.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { Wizard } from '$lib/layout';
-    import type { WizardStepsType } from '$lib/layout/wizard.svelte';
+    import type { WizardStepsType } from '$lib/layout/wizardWithSteps.svelte';
     import { sdk } from '$lib/stores/sdk';
     import { func } from '$routes/(console)/project-[project]/functions/function-[function]/store';
     import { wizard } from '$lib/stores/wizard';
