+++ src/routes/(console)/project-[project]/functions/create-function/wizard/createManual.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { ID, Runtime } from '@appwrite.io/console';
     import { Wizard } from '$lib/layout';
-    import type { WizardStepsType } from '$lib/layout/wizard.svelte';
+    import type { WizardStepsType } from '$lib/layout/wizardWithSteps.svelte';
     import { sdk } from '$lib/stores/sdk';
     import { wizard } from '$lib/stores/wizard';
     import { goto } from '$app/navigation';
