+++ src/routes/(console)/project-[project]/overview/platforms/createApple.svelte
@@ -3,7 +3,7 @@
     import { invalidate } from '$app/navigation';
     import { wizard } from '$lib/stores/wizard';
     import { createPlatform } from './wizard/store';
-    import type { WizardStepsType } from '$lib/layout/wizard.svelte';
+    import type { WizardStepsType } from '$lib/layout/wizardWithSteps.svelte';
     import Step1 from './wizard/apple/step1.svelte';
     import Step2 from './wizard/apple/step2.svelte';
     import Step3 from './wizard/apple/step3.svelte';
