+++ src/routes/(console)/project-[project]/overview/platforms/createReactNative.svelte
@@ -3,7 +3,7 @@
     import { invalidate } from '$app/navigation';
     import { wizard } from '$lib/stores/wizard';
     import { createPlatform } from './wizard/store';
-    import type { WizardStepsType } from '$lib/layout/wizard.svelte';
+    import type { WizardStepsType } from '$lib/layout/wizardWithSteps.svelte';
     import Step1 from './wizard/react-native/step1.svelte';
     import Step2 from './wizard/react-native/step2.svelte';
     import Step3 from './wizard/react-native/step3.svelte';
